webserver: Support per-site nginx apparmor override files
This creates local.d/usr.sbin.nginx.d/ in which override
files can be placed. This is the most convenient place,
as we can use it easily for sites (via vars), and other
roles can install their own overrides in there as well.

It's a good idea for upstreaming the apparmor service
file I think, as it allows plugin packages to provide
additional apparmor overrides in there as well.