mailserver: Replace postforward by a short rspamd lua snippet
This replaces the postforward binary by an rspamd lua snippet
that talks to postsrsd and rewrites the smtp from in the milter.

This needs a TODO to not do this for authenticated users, so we
only do it to incoming email.

This solves the problem that rspamd was called twice for each
forwarded message (once for the incoming, once for the forward),
while still preserving a proper envelope sender which we would not
have with just using postsrsd in a canonical sender map (IIRC).